requirements.txt
python-telegram-bot==20.7
vnstock>=3.1.0
pandas>=1.3
python-dotenv
